+++ ICD.Connect.Displays.Nec/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Reflection;
 [assembly: AssemblyTitle("ICD.Connect.Displays.Nec")]
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Displays.Nec")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2019")]
 [assembly: AssemblyVersion("12.1.0.0")]