+++ ICD.Connect.Displays.Panasonic/Devices/PanasonicClassicDisplay.cs
@@ -289,7 +289,7 @@ namespace ICD.Connect.Displays.Panasonic.Devices
 					break;
 
 				case "QAV":
-					string volume = ExtractParameter(response, 1);
+					string volume = ExtractParameter(response, 3);
 					Volume = int.Parse(volume);
 					IsMuted = false;
 					break;